Update csv_columns using itertools & setdefault
Instead of filling out values with a loop, use zip_longest with
fillvalue from the itertools module.  Also use dict.setdefault() to
create the initial list value for each key.  setdefault returns the
value of the key, which in this case is the list itself, which is why
you can use .append directly on it.