Update csv_columns nofill headers
The last solution would have filled headers with the missing value,
since headers and rows from the file were passed into zip_longest
together.  This solution uses zip_longest and fillvalue only with the
columns, then uses regular zip for the headers and columns.  In this
case, if the headers are shorter than the columns, it will truncate the
data to only show values for each header, and nothing extra.  This is
what would be expected in this case.  Also use headers variable from
definition instead of defining a new one, removing the else statement.