Update csv_columns to use csv.DictReader
Instead of using zip_longest and zip to iterate through the headers and
columns at the same time, just use csv.DictReader, which allows you to
iterate through row.items() for each line in the csv easily.  Also use
defaultdict from the collections module to create a dictionary that
defaults to lists.